Respond to PR #11 comments
